Track2 Migration: Dependency update (#1379)

+++ common/config/keys_tree_test.go
@@ -73,7 +73,7 @@ func (suite *keysTreeTestSuite) TestParseValue() {
 		{val: "65535", toType: reflect.Uint16, result: 65535},
 		{val: "4294967295", toType: reflect.Uint32, result: 4294967295},
 		{val: "18446744073709551615", toType: reflect.Uint64, result: uint64(18446744073709551615)},
-		{val: "6.24321908234", toType: reflect.Float32, result: 6.24321908234},
+		{val: "6.24321908234", toType: reflect.Float32, result: (float32)(6.24321908234)},
 		{val: "31247921747687123.123871293791263", toType: reflect.Float64, result: 31247921747687123.123871293791263},
 		{val: "6-8i", toType: reflect.Complex64, result: 6 - 8i},
 		{val: "2341241-910284i", toType: reflect.Complex128, result: 2341241 - 910284i},
